requirements.txt
asgiref==3.8.1
attrs==25.3.0
black==22.1.0
blue==0.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
Django==4.1
djangorestframework==3.14.0
drf-spectacular==0.28.0
flake8==4.0.1
idna==3.10
inflection==0.5.1
isort==6.0.1
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
mccabe==0.6.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
pathspec==0.12.1
platformdirs==4.3.8
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
stripe==12.3.0
tomli==2.2.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
